Some compiler warning fixes from Florent that got lost in the shuffle..

+++ include/pi-serial.h
@@ -14,9 +14,9 @@ extern "C" {
 			struct pi_sockaddr *addr, socklen_t addrlen));
 		int (*close) PI_ARGS((pi_socket_t *ps));
 		int (*changebaud) PI_ARGS((pi_socket_t *ps));
-		int (*write) PI_ARGS((pi_socket_t *ps,
+		ssize_t (*write) PI_ARGS((pi_socket_t *ps,
 			unsigned char *buf, size_t len, int flags));
-		int (*read) PI_ARGS((pi_socket_t *ps,
+		ssize_t (*read) PI_ARGS((pi_socket_t *ps,
 			unsigned char *buf, size_t len, int flags));
 		int (*poll) PI_ARGS((pi_socket_t *ps, int timeout));
 	};
